+++ hoss/utilities.py
@@ -13,11 +13,17 @@ from typing import Dict, Optional, Set, Tuple
 from urllib.parse import quote
 from uuid import uuid4
 
-from harmony_service_lib.exceptions import ForbiddenException, ServerException
+from harmony_service_lib.exceptions import (
+    ForbiddenException,
+    HarmonyException,
+    NoRetryException,
+    ServerException,
+)
 from harmony_service_lib.util import Config
 from harmony_service_lib.util import download as util_download
 
-from hoss.exceptions import UrlAccessFailed
+from hoss.exceptions import CustomNoRetryError, UrlAccessFailed
+from hoss.harmony_log_context import get_logger
 
 
 def get_file_mimetype(file_name: str) -> Tuple[Optional[str], Optional[str]]:
@@ -39,7 +45,6 @@ def get_opendap_nc4(
     url: str,
     required_variables: Set[str],
     output_dir: str,
-    logger: Logger,
     access_token: str,
     config: Config,
 ) -> str:
@@ -61,7 +66,6 @@ def get_opendap_nc4(
     downloaded_nc4 = download_url(
         netcdf4_url,
         output_dir,
-        logger,
         access_token=access_token,
         config=config,
         data=request_data,
@@ -99,7 +103,6 @@ def move_downloaded_nc4(output_dir: str, downloaded_file: str) -> str:
 def download_url(
     url: str,
     destination: str,
-    logger: Logger,
     access_token: str = None,
     config: Config = None,
     data=None,
@@ -116,14 +119,19 @@ def download_url(
     content from the URL.
 
     """
-    logger.info(f'Downloading: {url}')
+    get_logger().info(f'Downloading: {url}')
 
     if data is not None:
-        logger.info(f'POST request data: "{format_dictionary_string(data)}"')
+        get_logger().info(f'POST request data: "{format_dictionary_string(data)}"')
 
     try:
         response = util_download(
-            url, destination, logger, access_token=access_token, data=data, cfg=config
+            url,
+            destination,
+            get_logger(),
+            access_token=access_token,
+            data=data,
+            cfg=config,
         )
     except ForbiddenException as harmony_exception:
         raise UrlAccessFailed(url, 400) from harmony_exception
@@ -158,3 +166,22 @@ def get_value_or_default(value: Optional[float], default: float) -> float:
 
     """
     return value if value is not None else default
+
+
+def raise_from_hoss_exception(exception: Exception):
+    """Convert a HOSS exception to an appropriate Harmony exception.
+
+    Translates HOSS-specific exceptions into Harmony framework exceptions,
+    preserving the exception chain. CustomNoRetryError exceptions are converted
+    to NoRetryException, while all other exceptions become HarmonyException
+    instances.
+
+    """
+    if issubclass(type(exception), CustomNoRetryError):
+        ExceptionClass = NoRetryException
+    else:
+        ExceptionClass = HarmonyException
+
+    raise ExceptionClass(
+        'Subsetter failed with error: ' + str(exception)
+    ) from exception