+++ hoss/variable_utilities.py
@@ -6,20 +6,19 @@ varinfo.
 """
 
 import re
-from logging import Logger
 from typing import Set
 
 from harmony_service_lib.message import Variable as HarmonyVariable
 from varinfo import VarInfoFromNetCDF4
 
 from hoss.exceptions import InvalidVariableRequest
+from hoss.harmony_log_context import get_logger
 from hoss.utilities import format_variable_set_string
 
 
 def check_invalid_variable_request(
     requested_variables: Set[HarmonyVariable],
     varinfo: VarInfoFromNetCDF4,
-    logger: Logger,
 ) -> None:
     """Check if explicitly requested variables are listed as excluded in
     the varinfo configuration, and if so throw an exception listing the
@@ -34,7 +33,7 @@ def check_invalid_variable_request(
     # If no variables are requested, all variables will be returned and the
     # varinfo exclusions will automatically be applied.
     if not requested_variables:
-        logger.info(
+        get_logger().info(
             f'All variables are requested. The following variables will be excluded: {unprocessable_variables}'
         )
         return
@@ -50,7 +49,7 @@ def check_invalid_variable_request(
             format_variable_set_string(requested_unprocessable_variables)
         )
 
-    logger.info('No invalid variables are requested.')
+    get_logger().info('No invalid variables are requested.')
     return